Add files via upload
This is the documentation of my method. I may upload a clearer version later(when my printer is fixed).
I have also tested it on several algorithms, and it has been effective so far. I can't upload the tests yet because I want to print the algorithms I’m analyzing and attach them to the paper, but my printer broke down.